Segmentation fault

+++ p04-hash-table/lib/Dispersion/ModuleDispersion.h
@@ -21,7 +21,7 @@ class ModuleDispersion : public DispersionFunction<Key> {
   public:
     ModuleDispersion(unsigned int size) : DispersionFunction<Key>(size) {}
     unsigned operator()(const Key& key) const override {
-      return key % this->getTableSize();
+      return (static_cast<long>(key) % this->getTableSize());
     }
 };
